First of the Project.
This is the first of the project in a functional format. Lacks some of the things I might consider adding to it, but it does fire up and is functional. Creation of the window populates the peers list, but if load in a new character with it running the peers list doesn't populate the combo box. This is likely because DanNet hasn't connected that character yet and therefore they have no peers to add. The peers list doesn't automatically update when there is a change to the list of peers, something that can be handled later. I may consider changing some function names and such. So consider this a rough draft. 

Also not a fan of the ParseMacroData, but what other choice do I have.